Incorporate the suffix-only case into the general diffing
Since the original duplication checking only relied on product name checking,
this no longer allow the change in suffix that is now allowed by new
coronography association rules.  To incorporate the new case, a bit of
re-arrangement of code and diffing concepts were needed.
